New returns parsing errors; better documentation
While I really do not like forcing the user to check an error return
value from calling New, I think it's the most sensible solution, to
allow for errors to be returned when the log template line has a
noticable formatting error. I added a `MustCompile` method to the test
file. `MustCompile` is to `New` as `regexp.MustCompile` is to
`regex.Compile`. However, `gologs.MustCompile` remains in the test
file for now, so it is not yet part of the API.