work-in-progress to implement io.Writer
It would be quite beneficial to be able to use this library as an
io.Writer to pass along to other libraries for composition
functionality.

While initially the idea seemed simple there are some edge cases that
make it a bit more tricky than I originally assumed. The concept seems
valid, and I wanted to capture the work already partiallly completed.